# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Templarc",
...
    "contact": {
      "name": "Templarc"
    },
    "license": {
      "name": "Private"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/catalog/projects/{project_id}/templates": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "List templates as tree",
        "description": "Return the template hierarchy for a project as a nested tree. Only active templates are included. Nodes are ordered by sort_order then name within each level.",
        "operationId": "get_project_templates_catalog_projects__project_id__templates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateTreeNode"
                  },
                  "title": "Response Get Project Templates Catalog Projects  Project Id  Templates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/templates": {
      "get": {
        "tags": [
          "Templates"
        ],
        "summary": "List templates",
        "description": "Return templates, optionally filtered by `project_id`. Only active templates are returned by default; pass `active_only=false` to include soft-deleted ones.",
        "operationId": "list_templates_templates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateOut"
                  },
                  "title": "Response List Templates Templates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Templates"
        ],
        "summary": "Create template",
        "description": "Create a template record and write its initial `.j2` file to Git.\n\n- `parent_template_id` must belong to the same project if provided.\n- `git_path` is auto-generated as `{project.git_path}/{name}.j2` if omitted.\n- If `content` is empty, a minimal frontmatter stub is written.",
        "operationId": "create_template_templates_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/templates/{template_id}": {
      "get": {
        "tags": [
          "Templates"
        ],
        "summary": "Get template",
        "description": "Fetch a template record by ID.",
        "operationId": "get_template_templates__template_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Templates"
        ],
        "summary": "Update template",
        "description": "Update template metadata and, when `content` is provided, write a new Git commit with the updated `.j2` file.\n\nThe response always includes a `suggested_parameters` list \u2014 variable references parsed from the new body, annotated with `is_registered` to show which are already in the parameter registry.",
        "operationId": "update_template_templates__template_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Template Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateUpdateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Templates"
        ],
        "summary": "Delete template",
        "description": "Soft-delete a template by setting `is_active=False`. The `.j2` file is **not** removed from Git \u2014 history is preserved.",
        "operationId": "delete_template_templates__template_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/templates/{template_id}/variables": {
      "get": {
        "tags": [
          "Templates"
        ],
        "summary": "Get template variables",
        "description": "Parse the template's `.j2` file and return all Jinja2 variable references, each annotated with `is_registered` \u2014 whether a matching parameter exists in the registry (template-scope or project-scope).",
        "operationId": "get_template_variables_templates__template_id__variables_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VariableRefOut"
                  },
                  "title": "Response Get Template Variables Templates  Template Id  Variables Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/templates/{template_id}/inheritance-chain": {
      "get": {
        "tags": [
          "Templates"
        ],
        "summary": "Get inheritance chain",
        "description": "Return the full parent chain for a template, from the root ancestor down to the requested template (inclusive), ordered root-first.\n\nThis reflects the catalog hierarchy used for parameter inheritance (not Jinja2 `extends`).",
        "operationId": "get_inheritance_chain_templates__template_id__inheritance_chain_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InheritanceChainItem"
                  },
                  "title": "Response Get Inheritance Chain Templates  Template Id  Inheritance Chain Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/parameters": {
      "get": {
        "tags": [
          "Parameters"
        ],
        "summary": "List parameters",
        "description": "Return a paginated list of parameters. Filter by scope, project, template, organization, or a search string (name/label).",
        "operationId": "list_parameters_parameters_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ParameterScope"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by scope",
              "title": "Scope"
            },
            "description": "Filter by scope"
          },
          {
            "name": "organization_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Organization Id"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "template_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Template Id"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search in name or label",
              "title": "Search"
            },
            "description": "Search in name or label"
          },
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include soft-deleted parameters",
              "default": false,
              "title": "Include Inactive"
            },
            "description": "Include soft-deleted parameters"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ParameterOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Parameters"
        ],
        "summary": "Create parameter",
        "description": "Create a new parameter. Scoping rules are enforced:\n- `global`: name must start with `glob.`, `organization_id` required\n- `project`: name must start with `proj.`, `project_id` required\n- `template`: name must NOT start with `glob.` or `proj.`, `template_id` required",
        "operationId": "create_parameter_parameters_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParameterCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ParameterScope": {
        "type": "string",
        "enum": [
          "global",
          "project",
          "template"
        ],
        "title": "ParameterScope"
      },
...
      "TemplateOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "git_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Path"
          },
          "parent_template_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Template Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "name",
          "display_name",
          "description",
          "git_path",
          "parent_template_id",
          "is_active",
          "sort_order",
          "created_at",
          "updated_at"
        ],
        "title": "TemplateOut",
        "examples": [
          {
            "created_at": "2024-01-15T09:00:00Z",
            "description": "Cisco 891 router base configuration",
            "display_name": "Cisco 891",
            "git_path": "router_provisioning/cisco_891.j2",
            "id": 5,
            "is_active": true,
            "name": "cisco_891",
            "project_id": 1,
            "sort_order": 10,
            "updated_at": "2024-01-15T09:00:00Z"
          }
        ]
      },
      "TemplateTreeNode": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/TemplateTreeNode"
            },
            "type": "array",
            "title": "Children",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "display_name",
          "is_active",
          "sort_order"
        ],
        "title": "TemplateTreeNode"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}